Re-ordering of Kernel actions to vectorially sum particle displacements

+++ tests/test_advection.py
@@ -13,7 +13,6 @@ from parcels import (  # noqa
     AdvectionRK4,
     AdvectionRK4_3D,
     AdvectionRK45,
-    ErrorCode,
     Field,
     FieldSet,
     JITParticle,
@@ -24,11 +23,12 @@ from parcels import (  # noqa
     ParticleSetAOS,
     ParticleSetSOA,
     ScipyParticle,
-    StateCode,
+    StatusCode,
+    Variable,
     logger,
 )
 
-pset_modes = ['soa', 'aos']
+pset_modes = ['soa']
 ptype = {'scipy': ScipyParticle, 'jit': JITParticle}
 pset_type = {'soa': {'pset': ParticleSetSOA, 'pfile': ParticleFileSOA, 'kernel': KernelSOA},
              'aos': {'pset': ParticleSetAOS, 'pfile': ParticleFileAOS, 'kernel': KernelAOS}}
@@ -134,7 +134,7 @@ def test_advection_3D(pset_mode, mode, npart=11):
                                         depth=np.linspace(0, 1, npart))
     time = delta(hours=2).total_seconds()
     pset.execute(AdvectionRK4, runtime=time, dt=delta(seconds=30))
-    assert np.allclose(pset.depth*time, pset.lon, atol=1.e-1)
+    assert np.allclose(pset.depth*pset.time, pset.lon, atol=1.e-1)
 
 
 @pytest.mark.parametrize('pset_mode', pset_modes)
@@ -153,20 +153,25 @@ def test_advection_3D_outofbounds(pset_mode, mode, direction, wErrorThroughSurfa
     fieldset = FieldSet.from_data(data, dimensions, mesh='flat')
 
     def DeleteParticle(particle, fieldset, time):
-        particle.delete()
+        if particle.state == StatusCode.ErrorOutOfBounds or particle.state == StatusCode.ErrorThroughSurface:
+            particle.delete()
 
     def SubmergeParticle(particle, fieldset, time):
-        particle.depth = 0
-        AdvectionRK4(particle, fieldset, time)  # perform a 2D advection because vertical flow will always push up in this case
-        particle.time = time + particle.dt  # to not trigger kernels again, otherwise infinite loop
-        particle.set_state(StateCode.Success)
+        if particle.state == StatusCode.ErrorThroughSurface:
+            (u, v) = fieldset.UV[particle]
+            particle_dlon = u * particle.dt  # noqa
+            particle_dlat = v * particle.dt  # noqa
+            particle_ddepth = 0.  # noqa
+            particle.depth = 0
+            particle.state = StatusCode.Evaluate
 
-    recovery_dict = {ErrorCode.ErrorOutOfBounds: DeleteParticle}
+    kernels = [AdvectionRK4_3D]
     if wErrorThroughSurface:
-        recovery_dict[ErrorCode.ErrorThroughSurface] = SubmergeParticle
+        kernels.append(SubmergeParticle)
+    kernels.append(DeleteParticle)
 
     pset = pset_type[pset_mode]['pset'](fieldset=fieldset, pclass=ptype[mode], lon=0.5, lat=0.5, depth=0.9)
-    pset.execute(AdvectionRK4_3D, runtime=10., dt=1, recovery=recovery_dict)
+    pset.execute(kernels, runtime=11., dt=1)
 
     if direction == 'up' and wErrorThroughSurface:
         assert np.allclose(pset.lon[0], 0.6)
@@ -265,7 +270,7 @@ def test_length1dimensions(pset_mode, mode, u, v, w):
     pset = pset_type[pset_mode]['pset'](fieldset, pclass=ptype[mode], lon=x0, lat=y0, depth=z0)
     pfunc = AdvectionRK4 if w is None else AdvectionRK4_3D
     kernel = pset.Kernel(pfunc)
-    pset.execute(kernel, runtime=4)
+    pset.execute(kernel, runtime=5, dt=1)
 
     assert (len(pset.lon) == len([p.lon for p in pset]))
     assert ((np.array([p.lon - x0 for p in pset]) - 4 * u) < 1e-6).all()
@@ -312,11 +317,18 @@ def test_stationary_eddy(pset_mode, fieldset_stationary, mode, method, rtol, dif
         fieldset.add_constant('dres', 0.1)
     lon = np.linspace(12000, 21000, npart)
     lat = np.linspace(12500, 12500, npart)
-    pset = pset_type[pset_mode]['pset'](fieldset, pclass=ptype[mode], lon=lon, lat=lat)
+    dt = delta(minutes=3).total_seconds()
     endtime = delta(hours=6).total_seconds()
-    pset.execute(kernel[method], dt=delta(minutes=3), endtime=endtime)
-    exp_lon = [truth_stationary(x, y, endtime)[0] for x, y, in zip(lon, lat)]
-    exp_lat = [truth_stationary(x, y, endtime)[1] for x, y, in zip(lon, lat)]
+
+    class RK45Particles(ptype[mode]):
+        next_dt = Variable('next_dt', dtype=np.float32, initial=dt)
+
+    pclass = RK45Particles if method == 'RK45' else ptype[mode]
+    pset = pset_type[pset_mode]['pset'](fieldset, pclass=pclass, lon=lon, lat=lat)
+    pset.execute(pset.Kernel(kernel[method], delete_cfiles=False), dt=dt, endtime=endtime)
+
+    exp_lon = [truth_stationary(x, y, pset[0].time)[0] for x, y, in zip(lon, lat)]
+    exp_lat = [truth_stationary(x, y, pset[0].time)[1] for x, y, in zip(lon, lat)]
     assert np.allclose(pset.lon, exp_lon, rtol=rtol)
     assert np.allclose(pset.lat, exp_lat, rtol=rtol)
 
@@ -328,6 +340,7 @@ def test_stationary_eddy_vertical(pset_mode, mode, npart=1):
     lat = np.linspace(10000, 20000, npart)
     depth = np.linspace(12500, 12500, npart)
     endtime = delta(hours=6).total_seconds()
+    dt = delta(minutes=3).total_seconds()
 
     xdim = ydim = 100
     lon_data = np.linspace(0, 25000, xdim, dtype=np.float32)
@@ -342,9 +355,10 @@ def test_stationary_eddy_vertical(pset_mode, mode, npart=1):
     fieldset = FieldSet.from_data(data, dimensions, mesh='flat', transpose=True)
 
     pset = pset_type[pset_mode]['pset'](fieldset, pclass=ptype[mode], lon=lon, lat=lat, depth=depth)
-    pset.execute(AdvectionRK4_3D, dt=delta(minutes=3), endtime=endtime)
-    exp_lon = [truth_stationary(x, z, endtime)[0] for x, z, in zip(lon, depth)]
-    exp_depth = [truth_stationary(x, z, endtime)[1] for x, z, in zip(lon, depth)]
+    pset.execute(AdvectionRK4_3D, dt=dt, endtime=endtime)
+    exp_lon = [truth_stationary(x, z, pset[0].time)[0] for x, z, in zip(lon, depth)]
+    exp_depth = [truth_stationary(x, z, pset[0].time)[1] for x, z, in zip(lon, depth)]
+    print(pset, exp_lon)
     assert np.allclose(pset.lon, exp_lon, rtol=1e-5)
     assert np.allclose(pset.lat, lat, rtol=1e-5)
     assert np.allclose(pset.depth, exp_depth, rtol=1e-5)
@@ -353,9 +367,9 @@ def test_stationary_eddy_vertical(pset_mode, mode, npart=1):
     fieldset = FieldSet.from_data(data, dimensions, mesh='flat', transpose=True)
 
     pset = pset_type[pset_mode]['pset'](fieldset, pclass=ptype[mode], lon=lon, lat=lat, depth=depth)
-    pset.execute(AdvectionRK4_3D, dt=delta(minutes=3), endtime=endtime)
-    exp_depth = [truth_stationary(z, y, endtime)[0] for z, y, in zip(depth, lat)]
-    exp_lat = [truth_stationary(z, y, endtime)[1] for z, y, in zip(depth, lat)]
+    pset.execute(AdvectionRK4_3D, dt=dt, endtime=endtime)
+    exp_depth = [truth_stationary(z, y, pset[0].time)[0] for z, y, in zip(depth, lat)]
+    exp_lat = [truth_stationary(z, y, pset[0].time)[1] for z, y, in zip(depth, lat)]
     assert np.allclose(pset.lon, lon, rtol=1e-5)
     assert np.allclose(pset.lat, exp_lat, rtol=1e-5)
     assert np.allclose(pset.depth, exp_depth, rtol=1e-5)
@@ -399,11 +413,18 @@ def test_moving_eddy(pset_mode, fieldset_moving, mode, method, rtol, diffField, 
         fieldset.add_constant('dres', 0.1)
     lon = np.linspace(12000, 21000, npart)
     lat = np.linspace(12500, 12500, npart)
-    pset = pset_type[pset_mode]['pset'](fieldset, pclass=ptype[mode], lon=lon, lat=lat)
+    dt = delta(minutes=3).total_seconds()
     endtime = delta(hours=6).total_seconds()
-    pset.execute(kernel[method], dt=delta(minutes=3), endtime=endtime)
-    exp_lon = [truth_moving(x, y, endtime)[0] for x, y, in zip(lon, lat)]
-    exp_lat = [truth_moving(x, y, endtime)[1] for x, y, in zip(lon, lat)]
+
+    class RK45Particles(ptype[mode]):
+        next_dt = Variable('next_dt', dtype=np.float32, initial=dt)
+
+    pclass = RK45Particles if method == 'RK45' else ptype[mode]
+    pset = pset_type[pset_mode]['pset'](fieldset, pclass=pclass, lon=lon, lat=lat)
+    pset.execute(kernel[method], dt=dt, endtime=endtime)
+
+    exp_lon = [truth_moving(x, y, pset.time)[0] for x, y, in zip(lon, lat)]
+    exp_lat = [truth_moving(x, y, pset.time)[1] for x, y, in zip(lon, lat)]
     assert np.allclose(pset.lon, exp_lon, rtol=rtol)
     assert np.allclose(pset.lat, exp_lat, rtol=rtol)
 
@@ -459,11 +480,18 @@ def test_decaying_eddy(pset_mode, fieldset_decaying, mode, method, rtol, diffFie
         fieldset.add_constant('dres', 0.1)
     lon = np.linspace(12000, 21000, npart)
     lat = np.linspace(12500, 12500, npart)
-    pset = pset_type[pset_mode]['pset'](fieldset, pclass=ptype[mode], lon=lon, lat=lat)
+    dt = delta(minutes=3).total_seconds()
     endtime = delta(hours=6).total_seconds()
-    pset.execute(kernel[method], dt=delta(minutes=3), endtime=endtime)
-    exp_lon = [truth_decaying(x, y, endtime)[0] for x, y, in zip(lon, lat)]
-    exp_lat = [truth_decaying(x, y, endtime)[1] for x, y, in zip(lon, lat)]
+
+    class RK45Particles(ptype[mode]):
+        next_dt = Variable('next_dt', dtype=np.float32, initial=dt)
+
+    pclass = RK45Particles if method == 'RK45' else ptype[mode]
+    pset = pset_type[pset_mode]['pset'](fieldset, pclass=pclass, lon=lon, lat=lat)
+    pset.execute(kernel[method], dt=dt, endtime=endtime)
+
+    exp_lon = [truth_decaying(x, y, pset.time)[0] for x, y, in zip(lon, lat)]
+    exp_lat = [truth_decaying(x, y, pset.time)[1] for x, y, in zip(lon, lat)]
     assert np.allclose(pset.lon, exp_lon, rtol=rtol)
     assert np.allclose(pset.lat, exp_lat, rtol=rtol)
 
@@ -515,14 +543,14 @@ def test_uniform_analytical(pset_mode, mode, u, v, w, direction, tmpdir):
     outfile = pset.ParticleFile(name=outfile_path, outputdt=1, chunks=(1, 1))
     pset.execute(AdvectionAnalytical, runtime=4, dt=direction,
                  output_file=outfile)
-    assert np.abs(pset.lon - x0 - 4 * u * direction) < 1e-6
-    assert np.abs(pset.lat - y0 - 4 * v * direction) < 1e-6
+    assert np.abs(pset.lon - x0 - pset.time * u) < 1e-6
+    assert np.abs(pset.lat - y0 - pset.time * v) < 1e-6
     if w:
-        assert np.abs(pset.depth - z0 - 4 * w * direction) < 1e-4
+        assert np.abs(pset.depth - z0 - pset.time * w) < 1e-4
 
-    ds = xr.open_zarr(outfile_path, mask_and_scale=False)
-    times = ds['time'][:].values.astype('timedelta64[s]')[0]
-    timeref = direction * np.arange(0, 5).astype('timedelta64[s]')
+    ds = xr.open_zarr(outfile_path)
+    times = (direction*ds['time'][:]).values.astype('timedelta64[s]')[0]
+    timeref = np.arange(1, 5).astype('timedelta64[s]')
     assert np.allclose(times, timeref, atol=np.timedelta64(1, 'ms'))
     lons = ds['lon'][:].values
-    assert np.allclose(lons, x0+direction*u*np.arange(0, 5))
+    assert np.allclose(lons, x0+direction*u*np.arange(1, 5))
